refactor: Separated the code in genja-core-python/lib.rs into modules.
