+++ genja-core-python/python/genja_core/settings.py
@@ -1,0 +1,35 @@
+"""Python settings API for Genja.
+
+Import settings-facing helpers from this module instead of from ``genja_core``
+directly. The top-level package re-exports these names for compatibility, but
+``genja_core.settings`` is the primary public surface for:
+
+- ``Settings``
+- ``CoreConfig``
+- ``InventoryConfig``
+- ``OptionsConfig``
+- ``SSHConfig``
+- ``RunnerConfig``
+- ``LoggingConfig``
+"""
+
+from .genja_core import (
+    CoreConfig,
+    InventoryConfig,
+    LoggingConfig,
+    OptionsConfig,
+    RunnerConfig,
+    SSHConfig,
+    Settings,
+)
+
+
+__all__ = [
+    "Settings",
+    "CoreConfig",
+    "InventoryConfig",
+    "OptionsConfig",
+    "SSHConfig",
+    "RunnerConfig",
+    "LoggingConfig",
+]
